Add save menu item
