+++ src/utils/initIPCListeners.js
@@ -1,9 +1,44 @@
-import initSaveCopy from './initSaveCopy';
-import initConfirmClose from './initConfirmClose';
+import React from 'react';
+import { ipcRenderer } from 'electron';
+import { store } from '../ducks/store';
+import { actionCreators as protocolsActions } from '../ducks/modules/protocols';
+import { actionCreators as dialogActions } from '../ducks/modules/dialogs';
 
 const initIPCListeners = () => {
-  initConfirmClose();
-  initSaveCopy();
+  ipcRenderer.on('SAVE_COPY', (e, filePath) => {
+    store.dispatch(protocolsActions.saveCopy(filePath));
+  });
+
+  ipcRenderer.on('SAVE', () => {
+    store.dispatch(protocolsActions.saveAndExportProtocol());
+  });
+
+  ipcRenderer.on('CONFIRM_CLOSE', () => {
+    const state = store.getState();
+    const hasUnsavedChanges = state.session.lastChanged > state.session.lastSaved;
+
+    if (!hasUnsavedChanges) {
+      ipcRenderer.send('QUIT');
+      return;
+    }
+
+    store.dispatch(
+      dialogActions.openDialog({
+        type: 'Warning',
+        title: 'Unsaved changes',
+        message: (
+          <div>
+            Are you sure you want to exit the application?
+            <p><strong>Any unsaved changes will be lost!</strong></p>
+          </div>
+        ),
+        confirmLabel: 'Exit application',
+        onConfirm: () => {
+          ipcRenderer.send('QUIT');
+        },
+      }),
+    );
+  });
 };
 
 export default initIPCListeners;
